requirements.txt
streamlit
torch
torchvision
transformers
Pillow
numpy
opencv-python
requests
